move metrics section component to metrics folder
